added: seperate hook for fetching searched repo and focus in search input on inital visit
